Fix error calcul during requests
